Put index behind to load i18 early enough
